Revert "Completely working program!"
This reverts commit 19d463876223632aa0cf408e455dade8a081d02e.

+++ Trading Bot/Trading Bot/Logger/PredictionLogger.cs
@@ -20,7 +20,7 @@ namespace Trading_Bot.Logger
         /// Log prediction with time the prediciton should coorespond to.
         /// </summary>
         /// <param name="price">Predicted price.</param>
-        public void LogPrice(decimal price)
+        public void LogPrice(float price)
         {
             var currentTime = DateTime.Now;
             var timestamp = currentTime.AddSeconds(Configuration.Interval * Configuration.PredictionInterval);
